removed grading system, needed a conversion. Will do that in a later update. not a prio. Integegrated the tabs with the new state model. Not totally convinced. Could be better. Fine for now. Next up: create a double integer input

+++ app/lib/models/hold.dart
@@ -37,5 +37,5 @@ class Hold {
   final int repetitions;
   final int restBetweenRepetitions;
   final int hangTime;
-  final int addedWeight;
+  final double addedWeight;
 }
